+++ R/board.R
@@ -1,7 +1,11 @@
-new_board <- function(board, name, ...) {
+new_board <- function(board, name, cache, ...) {
+
+  if (is.null(cache)) stop("Please specify the 'cache' parameter, usually set to '~/.pins'.")
+
   board <- structure(list(
       board = board,
-      name = name
+      name = name,
+      cache = cache
     ),
     class = board)
 
@@ -14,20 +18,16 @@ new_board <- function(board, name, ...) {
 #'
 #' Connects to a board to activate RStudio's connection pane, when available.
 #'
-#' @param name The name of the board to activate.
+#' @param board The name of the board to activate.
+#' @param code The code being used to registere this board.
 #' @param ... Additional parameters required to initialize a particular board.
-#'
-#' @examples
-#'
-#' # define the storage location for all boards
-#' options(pins.path = tempdir())
 #'
 #' @keywords internal
 #' @export
-board_connect <- function(name, ...) {
-  board <- board_get(name)
+board_connect <- function(board, code, ...) {
+  board <- board_get(board)
 
-  ui_viewer_register(board)
+  ui_viewer_register(board, code)
 
   invisible(board)
 }
@@ -65,8 +65,8 @@ board_list <- function() {
 #' @param name The name of the board to use
 #'
 #' @export
-board_get <- function(name = NULL) {
-  if (is.null(name)) name <- getOption("pins.board", "local")
+board_get <- function(name) {
+  if (is.null(name)) name <- board_default()
 
   if (!name %in% board_list())
     stop("Board '", name, "' not a board, available boards: ", paste(board_list(), collapse = ", "))
@@ -76,41 +76,92 @@ board_get <- function(name = NULL) {
 
 #' Register Board
 #'
-#' Registers a board, useful to add sources to \code{pin_find()} or pin to remote
-#' boards with \code{pin()}.
+#' Registers a board, useful to find resources with \code{pin_find()} or pin to
+#' additional boards with \code{pin()}.
 #'
 #' @param board The name of the board to register.
 #' @param name An optional name to identify this board, defaults to the board name.
+#' @param cache The local folder to use as a cache, defaults to \code{"~/.pins"}.
 #' @param ... Additional parameters required to initialize a particular board.
+#'
+#' @details
+#'
+#' A board requires a local cache to avoid downloading files multiple times. It is
+#' recommended to not specify the \code{cache} parameter since it defaults to a well
+#' known \code{"~/.pins"} location. However, you are welcome to specify any other
+#' location for this cache or even a temp folder with \code{tempfile()}. Notice that,
+#' when using a temp folder, pins will be cleared when your R session restarts. The
+#' cache parameter can be also set with the \code{pins.path} option.
 #'
 #' @examples
 #' # create a new local board
-#' board_register("local", "other_board")
+#' board_register("local", "other_board", cache = tempfile())
 #'
 #' # create a Website board
-#' board_register("datatxt", name = "txtexample", url = "https://datatxt.org/data.txt")
+#' board_register("datatxt",
+#'                name = "txtexample",
+#'                url = "https://datatxt.org/data.txt",
+#'                cache = tempfile())
 #'
-#' @seealso \code{\link{board_register_github}}, \code{\link{board_register_kaggle}},
-#'   \code{\link{board_register_rsconnect}} and \code{\link{board_register_datatxt}}.
+#' @seealso \code{\link{board_register_local}}, \code{\link{board_register_github}},
+#'   \code{\link{board_register_kaggle}}, \code{\link{board_register_rsconnect}} and
+#'   \code{\link{board_register_datatxt}}.
 #'
 #' @export
-board_register <- function(board, name = board, ...) {
+board_register <- function(board, name = board, cache = "~/.pins", ...) {
   params <- list(...)
-  board <- new_board(board, name, ...)
+  board <- new_board(board, name, cache = cache, ...)
 
   board_registry_set(name, board)
 
-  if (!identical(params$connect, FALSE)) board_connect(name)
+  register_call <- board_register_code(board$name, name)
 
-  invisible(board)
+  if (!identical(params$connect, FALSE)) board_connect(name, register_call)
+
+  invisible(name)
+}
+
+# need to find the correct wrapper to support board_register_()
+board_register_code <- function(board, name) {
+  parent_idx <- 1
+  parent_call <- NULL
+  function_name <- NULL
+
+  while (parent_idx < length(sys.parents())) {
+    parent_func <- sys.function(sys.parent(parent_idx))
+    parent_call <- sys.call(sys.parent(parent_idx))
+    if (!is.function(parent_func) || !is.call(parent_call)) break;
+
+    this_parent_call <- tryCatch(match.call(definition = parent_func, call = parent_call), error = function(e) NULL)
+
+    if (is.null(this_parent_call)) break;
+    if (length(this_parent_call) < 1) break;
+
+    this_function_name <- deparse(this_parent_call[[1]])
+
+    if (!grepl("(^|::)board_register", this_function_name)) break;
+
+    parent_call <- this_parent_call
+    function_name <- this_function_name
+    parent_idx <- parent_idx + 1
+  }
+
+  header <- if (grepl("^pins::", function_name)) "" else "library(pins)\n"
+  if (is.null(parent_call)) {
+    paste0(header, "board_register(\"", board, "\", name = \"", name, "\")")
+  }
+  else {
+    main_call <- paste(deparse(parent_call, width.cutoff = 500), collapse = " ")
+    paste0(header, main_call)
+  }
 }
 
 #' Deregister Board
 #'
-#' Deregisters a board, useful to disable boards no longer in use. This operation
-#' removes all locally cached pins.
+#' Deregisters a board, useful to disable boards no longer in use.
 #'
 #' @param name An optional name to identify this board, defaults to the board name.
+#' @param ... Additional parameters required to deregister a particular board.
 #'
 #' @examples
 #'
@@ -124,15 +175,31 @@ board_register <- function(board, name = board, ...) {
 #' board_deregister("other_board")
 #'
 #' @export
-board_deregister <- function(name) {
+board_deregister <- function(name, ...) {
   if (!name %in% board_list()) stop("Board '", name, "' is not registered.")
 
   board <- board_get(name)
-  storage <- board_local_storage(board$name)
 
-  board_disconnect(name)
+  if (!identical(list(...)$disconnect, FALSE)) board_disconnect(name)
   board_registry_set(name, NULL)
-  unlink(storage, recursive = TRUE)
 
   invisible(NULL)
+}
+
+#' Default Board
+#'
+#' Retrieves the default board, which defaults to \code{"temp"} but can also be
+#' configured with the \code{pins.board} option.
+#'
+#' @examples
+#'
+#' # configure default board
+#' options(pind.board = "temp")
+#'
+#' # retrieve default board
+#' board_default()
+#'
+#' @export
+board_default <- function() {
+  getOption("pins.board", "temp")
 }