+++ R/board_rsconnect.R
@@ -36,17 +36,7 @@ board_initialize.rsconnect <- function(board, ...) {
   board
 }
 
-board_load.rsconnect <- function(board) {
-  board
-}
-
-board_persist.rsconnect <- function(board) {
-  board
-}
-
 board_pin_create.rsconnect <- function(board, path, name, metadata, ...) {
-  on.exit(board_connect(board$name))
-
   deps <- rsconnect_dependencies()
 
   temp_dir <- file.path(tempfile(), name)